+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   * added a missing setter for `shortName`
   * added `org.jenkins-ci.jpi` as alternative qualified plugin id for Gradle plugin portal inclusion
   * removed `WEB_APP_GROUP` constant from `org.jenkinsci.gradle.plugins.jpi.JpiPlugin`
+  * changed visibility of `org.jenkinsci.gradle.plugins.jpi.JpiPlugin.configureConfigurations` to private
   * `jpiDeployUser` and `jpiDeployPassword` properties from `org.jenkinsci.gradle.plugins.jpi.JpiExtension` were not
     used and have been removed
   * `org.jenkinsci.gradle.plugins.jpi.JpiPluginConvention` was not used and has been removed