+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.6.1 (unreleased)
+## 0.7.0 (2014-10-23)
 
   * updated Gradle to version 1.12
   * added support for `Plugin-Developers` manifest attribute